MOBILSDK-2803: refactor listeners
We always remove all the existing listeners when we add a new one, so
there's no reason to store these in an array; we just need to keep
around the latest listener. On the other hand, we can also set the class
itself as the listeners for the SDK just once on init; there's also no
need to wait to do this until we have a callback from the client.